Add OmitStrict type for when you don't want to be permissive. (#31)
I preferred the old behavior of `Omit` before it was made permissive, but I understand the use-cases for it. This PR re-introduces the original behavior under a different name.

The benefit that a stricter type has is primarily:

- Preventing typos.
- Allowing the compiler to pick up on rename refactors automatically.

Currently, permissive `Omit` acts as a "barrier" that prevents rename refactors from passing through, which means that any such refactor generates whole bunches of errors that have to be manually fixed. If the field in question is optional, this can actually introduce bugs.

I'm open to bike-shedding the name, but I'd like to have both types available so I don't have to pin to an old version of the library or re-declare my own strict variants!